PC-Pair Version 1.0 Complete
The main script is hard coded with five presets for pc builds at varying price ranges and scrapes Newegg and MicroCenter for the best available prices of each part in the selected preset.  It outputs a formatted list of each component's name, a link to the product page, and cheapest price.